Move elevator pid from stub to subsystem
Move software pid from elevator stub to elevator subsystem.  This will let us
use the same pid in simulation and on real robot/HW.

Add some commands that set the elevator to known positions based on reef.  The
constants are just guesses for now and will need to be tuned/set later.

Put some subystems and command scheduler in network tables so we can view
them in dashboard to see what commands are executing.